✨ feat: reset.tsx 구조 분리
